+++ js/gui/kijs.gui.Tree.js
@@ -843,6 +843,9 @@ kijs.gui.Tree = class kijs_gui_Tree extends kijs.gui.DataView {
                 };
                 newEl.ddSource.on('drop', this.#onElementSourceDrop, this);
             }
+            if (!kijs.isEmpty(newEl.ddSource)) {
+                newEl.ddSource.on('dragStart', this.#onElementSourceDragStart, this);
+            }
 
             if (this._elementDdTargetConfig) {
                 newEl.ddTarget = this._elementDdTargetConfig;
@@ -971,6 +974,23 @@ kijs.gui.Tree = class kijs_gui_Tree extends kijs.gui.DataView {
 
     // PRIVATE
     // LISTENERS
+    // overwrite
+    #onElementSourceDragStart(e) {
+        if (!this.disabled && !e.source.ownerEl.disabled) {
+            // Falls nicht selektiert: selektieren
+            if (!e.source.ownerEl.selected) {
+                this.current = e.source.ownerEl;
+                if (this._focusable) {
+                    e.source.ownerEl.focus();
+                }
+
+                let isShiftPress = false;
+                let isCtrlPress = false;
+                this._selectEl(this._currentEl, isShiftPress, isCtrlPress);
+            }
+        }
+    }
+
     // overwrite
     #onElementSourceDrop(e) {
         let dataRows = [];